requirements.txt
amqp==1.4.9
anyjson==0.3.3
awesome-slugify==1.6.5
billiard==3.3.0.23
blessings==1.6
blinker==1.4
boto3==1.3.0
botocore==1.4.5
celery==3.1.23
#cffi==1.3.1
cycler==0.10.0
dask==0.8.2
decorator==4.0.9
#dlib==18.17.100
docopt==0.4.0
docutils==0.12
fake-factory==0.6.0
Flask==0.10.1
Flask-Script==2.0.5
futures==2.2.0
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.7.1
kombu==3.0.35
MarkupSafe==0.23
#matplotlib==1.5.1
MongoFrames==1.2.4
networkx==1.11
nose==1.3.7
numpy==1.11.0
Pillow==5.0.0
py==1.4.31
pycparser==2.14
pymongo==3.3.0
pyparsing==2.1.1
pytest==2.8.7
pytest-flask==0.10.0
python-dateutil==2.5.3
pytz==2016.10
raven==5.5.0
regex==2016.3.2
requests==2.7.0
#scikit-image==0.12.3
#scipy==0.18.1
shortuuid==0.4.3
six==1.10.0
toolz==0.7.4
Unidecode==0.4.19
Werkzeug==0.10.4
WTForms==2.0.2
